create a entrypoint to domain with a reductor
