+++ src/main/scala/training/entrypoint/ShopReductor.scala
@@ -1,0 +1,20 @@
+package training.entrypoint
+
+import training.domain.Shop
+
+trait ShopReductor {
+  def createShop(
+                  businessName: String,
+                  name: String,
+                  activityId: Int,
+                  stratumId: Int,
+                  address: String,
+                  phoneNumber: String,
+                  email: String,
+                  website: String,
+                  shopTypeId: Int,
+                  position: (Long, Long)
+                ): Shop
+
+  def all(): List[Shop]
+}
